Adding customer to swagger and add implementation
