+++ src/backend/quarkus-management-api/src/main/java/com/nttdata/quarkus/management/api/controller/ManagementApiImpl.java
@@ -1,14 +1,18 @@
 package com.nttdata.quarkus.management.api.controller;
 
 import com.nttdata.quarkus.management.api.model.database.Contacts;
+import com.nttdata.quarkus.management.api.model.database.Customers;
 import com.nttdata.quarkus.management.api.openapi.ManagementApi;
 import com.nttdata.quarkus.management.api.openapi.model.Contact;
+import com.nttdata.quarkus.management.api.openapi.model.Customer;
 import com.nttdata.quarkus.management.api.service.ContactsService;
+import com.nttdata.quarkus.management.api.service.CustomersService;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
 import java.math.BigDecimal;
@@ -19,6 +23,9 @@ public class ManagementApiImpl implements ManagementApi {
 
     @Inject
     ContactsService contactsService;
+
+    @Inject
+    CustomersService customersService;
 
     @Override
     @Transactional
@@ -83,6 +90,71 @@ public class ManagementApiImpl implements ManagementApi {
         contacts.setPhone(contact.getPhoneNumber());
 
         return contacts;
+
+    }
+
+    @Override
+    @Transactional
+    public Response addCustomer(@Valid Customer customer, SecurityContext securityContext) {
+        System.out.println("Hello we have entered the add method");
+        if (customer.getCustomerId() != null) {
+            throw new WebApplicationException("Id not required.", 422);
+        }
+        return Response.ok(mapCustomer(customersService.addCustomer(mapCustomers(customer)))).status(200).build();
+    }
+
+    @Override
+    @Transactional
+    public Response deleteCustomer(BigDecimal customerId, SecurityContext securityContext) {
+        customersService.deleteCustomer(customerId.toBigInteger());
+        return Response.status(204).build();
+    }
+
+    @Override
+    @Transactional
+    public Response getCustomer(BigDecimal customerId, SecurityContext securityContext) {
+        return Response.ok(mapCustomer(customersService.getCustomer(customerId.toBigInteger()))).build();
+    }
+
+    @Override
+    @Transactional
+    public Response getCustomers(SecurityContext securityContext) {
+        return Response.ok(
+                customersService.getCustomers().stream()
+                .map(this::mapCustomer)
+                .collect(Collectors.toList())).build();
+    }
+
+    @Override
+    @Transactional
+    public Response updateCustomer(@Valid Customer customer, SecurityContext securityContext) {
+
+        return Response.ok(mapCustomer(customersService.updateCustomer(mapCustomers(customer)))).status(200).build();
+
+    }
+
+    private Customers mapCustomers(Customer customer) {
+
+        Customers customers = new Customers();
+        customers.setCustomerId((customer.getCustomerId() != null) ? customer.getCustomerId().toBigInteger() : null);
+        customers.setName(customer.getFullName());
+        customers.setAddress(customer.getAddress());
+        customers.setWebsite(customer.getWebsite());
+        customers.setCreditLimit(customer.getCreditLimit());
+        return customers;
+
+    }
+
+    private Customer mapCustomer(Customers customers) {
+
+        System.out.println("Hello we are mapping a api customer to a database customer");
+        Customer customer = new Customer();
+        customer.setCustomerId(BigDecimal.valueOf(customers.getCustomerId().longValue()));
+        customer.setFullName(customers.getName());
+        customer.setAddress(customers.getAddress());
+        customer.setWebsite(customers.getWebsite());
+        customer.setCreditLimit(customers.getCreditLimit());
+        return customer;
 
     }
 }
